Checkpoint - LEDs working now

+++ src/Renderer.cpp
@@ -1,36 +1,4 @@
-// TODO Remove?
-//#include <functional>
-//
 #include "Renderer.h"
-
-//void Renderer::loop(int numPixels, std::function<void(int, Color)> setPixel) {
-//  // If there's no model then there's nothing to do
-//  if (model == NULL) {
-//    return;
-//  }
-//
-//  // Update the current state of the model to match the current time.
-//  unsigned long const nowMS = millis();
-//  float const timeStamp = (nowMS - startTimeMS) / 1000.0;
-//  model->update(timeStamp);
-//
-//  // Now apply the model to each LED position.
-//  for (int i = 0; i < numPixels; i++) {
-//    // Get the color from the model
-//    float const pos = ((float)i) / (numPixels - 1);
-//    Color const color = model->render(pos);
-//
-//    // Apply gamma correction
-//    uint16_t const red = Colors::getRed(color);
-//    uint16_t const green = Colors::getGreen(color);
-//    uint16_t const blue = Colors::getBlue(color);
-//    Color const correctedColor = Colors::makeColor(
-//      (uint8_t)(red*red/255), (uint8_t)(green*green/255), (uint8_t)(blue*blue/255));
-//
-//    // Set the pixel on the light strip
-//    setPixel(i, correctedColor);
-//  }
-//}
 
 // TODO Doc
 Esp8266_NeoPixelBus_Renderer::Esp8266_NeoPixelBus_Renderer(int pixelsPin, int pixelsCount) :
@@ -39,9 +7,6 @@ Esp8266_NeoPixelBus_Renderer::Esp8266_NeoPixelBus_Renderer(int pixelsPin, int pi
         _pixelsCount(pixelsCount),
         _strip(pixelsCount, pixelsPin) {
     _strip.Begin();
-    // TODO Try brightness NeoPixelBus class
-//    _strip.SetBrightness(255);
-
     _strip.Show();  // Initialize all pixels to 'off'
 }
 
@@ -60,8 +25,7 @@ void Esp8266_NeoPixelBus_Renderer::render() {
 
     // Set the color of each pixel
     for (auto i = 0; i < _pixelsCount; ++i) {
-        // TODO Do I need to cast to force pos to be a float?
-        float pos = i / (_pixelsCount - 1);
+        float pos = (float)i / float(_pixelsCount - 1);
         auto color = _model->render(pos);
         RgbColor pixelColor(Colors::getRed(color), Colors::getGreen(color), Colors::getBlue(color));
         _strip.SetPixelColor(i, pixelColor);
